requirements.txt
Flask
openai==0.28
gunicorn
python-dotenv==1.0.1